requirements.txt
Django==1.7
django-allauth==0.25.2
django-appconf==1.0.1
django-embed-video==1.1.0
django-geoposition==0.2.3
django-image-cropping==1.0.3
django-model-utils==2.5
django-multiselectfield==0.1.3
django-notifications==0.1.dev0
django-notifications-hq==1.2
django-social-auth==0.7.28
django-timedeltafield==0.7.10
djangorestframework==3.3.3
easy-thumbnails==2.3
gunicorn==19.6.0
httplib2==0.9.2
ipgetter==0.6
jsonfield==1.0.3
notifications==0.3.2
oauth2==1.9.0.post1
oauthlib==1.1.2
Pillow==3.2.0
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
wheel==0.24.0
whitenoise==3.2
